Separate the storer in the file structure
